Add updated specs to build Firecracker
This includes edits to musl and rust to build Linux musl targets,
since glibc is not officially supported by Firecracker.  A handful
of added crates are just updated versions of packages already in
Fedora.  Crates that are not on the latest update track have the
major/minor suffix in the name.

The firecracker package itself bundles two crates that are forked
versions of existing crates.  It is currently limited to building
for x86_64, but aarch64 is supported upstream and may be added
later.  Rust's musl target is using its default behavior of static
linking, so the Firecracker binaries are standalone executables.