Provide rust-srpm-macros in RHEL
rust-srpm-macros is a dependency of redhat-rpm-config so that Rust-based
SRPMs can rely on these particular macros.  In RHEL 9 and earlier,
rust-srpm-macros was a separate source package which was included in
RHEL.  Since then, it became a subpackage of rust-packaging in Fedora.

In order to not pull in any part of rust-packaging into RHEL 10, thereby
allowing it to be maintained separately in EPEL, those macros are
provided here instead.