Fix the AArch64 build
AArch64 musl always requires compiler builtins for quad math, but
the compiler-builtins crate doesn't implement them in Rust.  The
only way to get them from the build system is to take the C object
files from the bundled compiler-rt source.  Rust's build system can
only enable this globally for all targets, and it's disabled in the
Fedora package.

So this uses that setting only when AArch64 musl is enabled, which
deviates all Rust targets from Fedora's configuration.  Since this
change breaks compatibility anway, it also disables the extra Rust
targets for AArch64 to save build time (and since at least the UEFI
one doesn't work with the compiler-rt object files).  This Rust RPM
for AArch64 really shouldn't be used outside building Firecracker.

If the compiler-builtins crate ever implements the quad functions
in Rust, all of this can be unified with the Fedora build again.