Move the target list definition after the rust_triple function
This has no functional effect, it is just to be nice to downstreams
for adding additional ABIs to the target list with minimal changes.
It makes macros parametric to pass parameters to the Lua function,
but this is still unused in this spec.

Also clean up the ppc64 architecture from the list since it's not
being build by ExclusiveArch anymore.

[skip changelog]