Update to version 1.12.1
This adds a new conditional to toggle building cpu-template-helper,
but it defaults to being enabled in Fedora builds for no change in
behavior.  It will disable itself by default when building for the
riscv64 architecture to allow testing with the Fedora RISC-V SIG.

Note that Amazon currently does not want to support official RISC-V
releases, so ExclusiveArch still excludes it.  The source archive
will have to be pointed at a riscv64 branch/commit, e.g. the most
recent is 141424bdeb1b31723b2290f47a319723b6e59049 in GitHub.