Update to version 1.12.1
This includes the most recent riscv64 branch for convenience.  It
won't affect patch fuzzing or even include the patch in the SRPM
unless it's built for riscv64.  It might be dropped in the future
if things get too out of sync, since Amazon does not want to merge
the change and officially support the platform.